feat(license): add the `@license` tag
The `@license` tag in the docstring annotates the license associated with
that object. Any text on the same line as `@license` will be recorded as
the license name. All text following until the end of the docstring or
the next tag is the optional license text. If more than one `@license`
tag is used, only the last one in the docstring will get recorded. Using
`@license` with no text will get ignored and return `None`.

Resolves: #18