feat(namespace): add the `@namespace` tag
The `@namespace` tag indicates that an object creates a namespace for its
members to help organize your project's documentation. The tag must have
a name that meets valid Python naming criteria. It may also contain an
optional description of the namespace. If the tags are omitted or
included but left blank, it will return `None`. Because namespaces are
distinct entities, it does not matter in which docstring you define them.
Therefore, you may define as many namespaces as needed in a docstring.

Additionally, while writing the tests for `namespace`, I found a bug in
the description parsing function. This commit has the fix to that bug
inseperably tied into it.

Resolves: #22, Resolves: #24