feat(memberof): add the `@memberof` tag
The `@memberof` tag indicates that an object belongs to a namespace or
module, even if that is not where the parser ultimately found it. This
helps organize your project's documentation. This is especially useful
when encapsulating functions, classes, or variables in other logically
separated files. Multiple `@memberof` tags can be used on a single object,
and this will cause the object to get duplicate documentation in the final
results. The tag specifies only a name that meets valid Python naming
criteria. If the tags are omitted or included but left blank, it will
return `None`. Otherwise, it returns a list of string names.

Resolves: #23