refactor(core): move the bulk of plugin code to `load_python_module`
The order that both plugins and templates will try to resolve the path
is the same for both. Reorganizing this allows easier expansion on the
next feature implementation, the template loading. This also removes what
turned out to be a redundant filter `template_path_before_loading`.

Numerous additional updates to improve console logging output messages.