fix(parser): handle errors when getting source
The parser was throwing an error when using `inspect.getsourcefile()` in
some cases. I could not track down why exactly it was doing that, but it
was during edge cases that was not the main purpose. This is a bandaid
that handles the issue without crashing the script.

Additionally, in the same vein I refactored the linestart code. This was
part of the same issue.