test(core): fix two broken tests
The default test broke because I included a config file. Therefore,
comparing to stock configuration was no longer a valid test assumption.
Additionally, a typo in a test name was preventing it from ever running
in the first place. Changing the test name resolved the issue with no
further problems.