feat(core): improve the actions and filters interface
The actions and filters reset their `.doing` parameter to nothing at the
conclusion of running the script now. This will allow users to see if
a hook has not only been done, but which one (if any) is currently in
action.

It also simplifies the console output for the filters to save a printed
line if the input is the exact same as the output.